requirements.txt
python-binance==1.0.19
pyspark==3.2.0
hdfs==2.7.2
pyarrow
pandas==1.5.3